+++ assets/tiles/kenney-sketch-desert.tsx
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.11.0" name="kenney-sketch-desert" tilewidth="256" tileheight="352" tilecount="5" columns="0">
+ <grid orientation="isometric" width="1" height="1"/>
+ <tile id="0">
+  <image source="kenney-sketch-desert/building_dark_sides_windows_E.png" width="256" height="352"/>
+ </tile>
+ <tile id="1">
+  <image source="kenney-sketch-desert/dirt_center_E.png" width="256" height="352"/>
+ </tile>
+ <tile id="2">
+  <image source="kenney-sketch-desert/grass_center_S.png" width="256" height="352"/>
+ </tile>
+ <tile id="3">
+  <image source="kenney-sketch-desert/stairs_full_N.png" width="256" height="352"/>
+ </tile>
+ <tile id="4">
+  <image source="kenney-sketch-desert/trees_N.png" width="256" height="352"/>
+  <objectgroup draworder="index" id="2">
+   <object id="3" x="-112" y="-372" width="274" height="234"/>
+  </objectgroup>
+ </tile>
+</tileset>