+++ assets/tiles/orthogonal_2.tsx
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.10.2" name="orthogonal_2" tilewidth="32" tileheight="32" tilecount="7" columns="0">
+ <grid orientation="orthogonal" width="1" height="1"/>
+ <tile id="0">
+  <image width="32" height="32" source="orthogonal/tile0.png"/>
+ </tile>
+ <tile id="1">
+  <image width="32" height="32" source="orthogonal/tile1.png"/>
+ </tile>
+ <tile id="2">
+  <image width="32" height="32" source="orthogonal/tile2.png"/>
+ </tile>
+ <tile id="3">
+  <image width="32" height="32" source="orthogonal/tile3.png"/>
+ </tile>
+ <tile id="4">
+  <image width="32" height="32" source="orthogonal/tile4.png"/>
+ </tile>
+ <tile id="5">
+  <image width="32" height="32" source="orthogonal/tile5.png"/>
+ </tile>
+ <tile id="6">
+  <image width="32" height="32" source="orthogonal/tile6.png"/>
+ </tile>
+</tileset>